// src/types/fire.ts
export interface FirePlacement {
  horizontalPosition: number;
  verticalPosition: number;
  flickerPhase: number;
}

export interface Ember {
  horizontalPosition: number;
  verticalPosition: number;
  velocityHorizontal: number;
  velocityVertical: number;
  secondsRemaining: number;
  totalSeconds: number;
}